+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+0.4.36
++++++
+* Added --uptime-sla for paid service
+
+0.4.35
++++++
+* Added support for creation time node labels
+
+0.4.34
++++++
+* Remove preview flag for private cluster feature.
+
 0.4.33
 +++++
 * Adding az aks get-credentials --context argument